requirements.txt
beautifulsoup4==4.4.1
boto==2.38.0
Django==1.9.1
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-htmlmin==0.8.0
django-storages-redux==1.3.1
django-widget-tweaks==1.4.1
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
html5lib==0.9999999
htmlmin==0.1.10
httpie==0.9.3
Pillow==3.1.0
psycopg2==2.6.1
Pygments==2.1
PyJWT==1.4.0
python-memcached==1.57
requests==2.9.1
six==1.10.0
sqlparse==0.1.18
wheel==0.24.0
